+++ src/main/java/de/carstorewebservice/CarstorewebserviceApplication.java
@@ -5,18 +5,11 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @SpringBootApplication
 public class CarstorewebserviceApplication {
-
-
-	@GetMapping("/registration")
-	public String reg(){
-		return "Ich regestriere dich!";
-	}
 
 	public static void main(String[] args) {
 		SpringApplication.run(CarstorewebserviceApplication.class, args);
