--- src/main/java/de/carstorewebservice/sql/repository/UserRepository.java
@@ -1,17 +1,0 @@
-package de.carstorewebservice.sql.repository;
-
-
-import de.carstorewebservice.sql.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    default List<User> findByLastName(String lastName) {
-        return null;
-    }
-
-    User findById(long id);
-}
-
